Created 'include/zarks/io' directory Moved relevant binary functions into the new location
