+++ include/zarks/io/binary.h
